+++ main/Sapphire/Utilities/Building/nanodeca.py
@@ -140,7 +140,6 @@ class Nanodeca:
          for i in range(len(dist)):
              new_dist.append(round(dist[i],3))
          values, counts = np.unique(new_dist, return_counts=True)
-         freq = np.dstack((values,counts)).reshape(-1,2)
          sorted_nano = nanoparticle[np.argsort(new_dist)]